feat: add db model
