+++ core/db.py
@@ -5,12 +5,13 @@ from typing import TYPE_CHECKING, Dict
 
 if TYPE_CHECKING:
     from . import LIPOIC
-    from .models import BaseModel, DevMember, Dvc
+    from .models import BaseModel, DevMember, Dvc, Member
 
 
 class DB:
     DevMember: 'DevMember'
     Dvc: 'Dvc'
+    Member: 'Member'
 
     def __init__(self, bot: 'LIPOIC'):
         self.bot = bot
